refacoted typography into sections
