+++ src/components/foundation/Typography.md
@@ -1,6 +1,27 @@
-Placeholder for documentation
+### Headings
 
 ```jsx
-import TypographyExample from './TypographyExample'
-;<TypographyExample />
+import HeadingExample from './HeadingExample'
+;<HeadingExample />
+```
+
+### Headings in a Dark Section
+
+```jsx
+import DarkHeadingExample from './DarkHeadingExample'
+;<DarkHeadingExample />
+```
+
+### Paragraph
+
+```jsx
+import BodyExample from './BodyExample'
+;<BodyExample />
+```
+
+### Links
+
+```jsx
+import LinksExample from './LinksExample'
+;<LinksExample />
 ```
